Added client module for authorization.
